Exempt ticket channels from link moderation
